% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_grid_adapt.R
\name{adapt2_grid}
\alias{adapt2_grid}
\title{Adaptive grid estimation of posteriors}
\usage{
adapt2_grid(params, ranges, priors, n_start, range_fixed, mix_grid, rho, X,
  max_post_per_bin, end_bin_thresh = 1e-08)
}
\description{
Workhorse for posterior adaptive grid approximation. Called from cause_grid_adapt.
}
