% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_download_example_gwas_data.R
\name{cause_download_example_gwas_data}
\alias{cause_download_example_gwas_data}
\title{Download GWAS data to use in examples}
\usage{
cause_download_example_gwas_data()
}
\description{
Download GWAS data to use in examples
}
\details{
This function downloads Snakemake summary statistics for LDL cholesterol (Willer et al 2013 PMID 24097068),
coronary artery disease (van der Harst et al 2017 PMID 29212778), and asthma (Demenais et al 2018 PMID 29273806) as well as a csv file that can be
used with the CAUSE pipeline.
}
