% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_grid_adapt.R
\name{cause_grid_adapt}
\alias{cause_grid_adapt}
\title{CAUSE posterior density estimation}
\usage{
cause_grid_adapt(X, param_ests, params = c("q", "gamma", "eta"),
  priors = list(function(q) {     dbeta(q, 1, 10) }, function(b) {    
  dnorm(b, 0, 0.6) }, function(b) {     dnorm(b, 0, 0.6) }),
  n_start = c(10, 20, 20), max_post_per_bin = 0.001)
}
\arguments{
\item{X}{An object of class cause_data}

\item{param_ests}{param_ests Object of class cause_params}

\item{params}{Vector of parameter names. Should be some subest of  c("q", "gamma", "eta") given in any order.}

\item{priors}{List of prior functions corresponding to parameters in params. Each element should be a
funciton that takes one argument and returns a log likelihood.}

\item{n_start}{Vector of length length(params) giving the number of starting bins for each dimension}

\item{max_post_per_bin}{Maximum posterior probability for each grid cube.
This parameter controls the fineness of the approximation.}
}
\value{
An object of class cause_post.
}
\description{
Uses an adaptive grid approximation to estimate the posterior distribution of q, gamma, and eta or a subset of those
}
