% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_cause_params_future.R
\name{est_cause_params_future}
\alias{est_cause_params_future}
\title{Estimate CAUSE Nuisance Parameters}
\usage{
est_cause_params_future(X, variants, sigma_g, qalpha = 1, qbeta = 10,
  optmethod = c("mixSQP", "mixIP"), sigma_g_pval = 0.001,
  null_wt = 10, max_candidates = Inf)
}
\arguments{
\item{X}{An object of class cause_data containing data for the two traits.}

\item{variants}{A vector of variants to include. This list should be approximately LD
pruned and include variants genome wide.}
}
\value{
An object of class cause_params
}
\description{
Estimates bivariate distribution of summary statistics and rho, the
correlation between summary statistics due to overlapping samples or population structure.
}
