% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_gamma_prior.R
\name{eta_gamma_prior}
\alias{eta_gamma_prior}
\title{eta/gamma prior}
\usage{
eta_gamma_prior(X, variants, prob = 0.05)
}
\arguments{
\item{X}{An object of class cause_data containing data for the two traits.}

\item{variants}{A vector of variants to include in the analysis.}

\item{prob}{See Details}
}
\value{
A list with items conf, full, elpd, summary, and plot.
}
\description{
Choose prior variance for eta and gamma empirically from data
}
\details{
The function will return a variance, sigma, such that
P(abs(z) > k) = prob where z is a N(0, sigma) random variable, k = max(abs(beta_hat_2/beta_hat_1))
and prob is specified in the arguments of the function.
}
