% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_format.R
\name{gwas_format}
\alias{gwas_format}
\title{Convert GWAS summary statistics to a standard format}
\usage{
gwas_format(X, snp, beta_hat, se, A1, A2, chrom, pos, p_value, sample_size,
  output_file, compute_pval = TRUE)
}
\arguments{
\item{X}{data.frame}

\item{snp}{Column name containing SNP ID}

\item{beta_hat}{Column name containing effect estimate}

\item{se}{Column name containing standard error of beta_hat}

\item{A1}{Column name containing effect allele}

\item{A2}{Column name containing other allele}

\item{chrom}{Chromosome column (optional)}

\item{pos}{Position column (optional)}

\item{p_value}{p-value column (optional)}

\item{sample_size}{Sample size column (optional) or an integer}

\item{output_file}{File to write out formatted data. If missing formatted data will be returned.}

\item{compute_pval}{Logical, compute the p-value using a normal approximation if missing? Defaults to TRUE.}
}
\value{
An object of class cause_data and data.frame.
}
\description{
Format GWAS summary statistics for CAUSE
}
\details{
This function will try to merge data sets X1 and X2 on the specified columns. Where
necessary, it will flip the sign of effects so that the effect allele is the same in both
data sets. It will remove variants with ambiguous alleles or where the alleles (G/C or A/T) or
with alleles that do not match between data sets (e.g A/G in one data set and A/C in the other).
It will not remove variants that are simply strand flipped between the two data sets (e. g. A/C in one data set, T/G in the other).
}
