% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample_elpd_loo.R
\name{in_sample_elpd_loo}
\alias{in_sample_elpd_loo}
\title{Estimate delta elpd for pairs of models}
\usage{
in_sample_elpd_loo(X, fits, variants, nsamps = 1000)
}
\arguments{
\item{X}{Data}

\item{fits}{List of models of class cause_post.}

\item{variants}{Optional vector of SNPs to use for ELPD calculation}

\item{nsamps}{Number of samples to take from the posterior}
}
\value{
A data frame with one row per pair of models giving the estimated
difference in elpds and standard error of that difference.
}
\description{
Estimate delta elpd for pairs of models
}
