% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_prune.R
\name{ld_prune}
\alias{ld_prune}
\title{Prune variants for LD}
\usage{
ld_prune(variants, ld, total_ld_variants, pval_cols, pval_thresh,
  r2_thresh = 0.1, variant_name = "snp")
}
\arguments{
\item{variants}{Either a vector of SNP names or a data.frame. If a data.frame, provide pval_cols
containing the names of pvalue columns to use for pruning. Results will contain one list per p-value
column. If variants is a vector, LD pruning will be random.}
}
\description{
Get an LD pruned list of variants optionally prioritizing variants with low p-values
}
