% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik}
\alias{loglik}
\title{Calculate likelihood -- version 7}
\usage{
loglik(rho, g, gp, q, tsigma1, tsigma2, tpi, tbeta_hat_1, tbeta_hat_2,
  tseb1, tseb2)
}
\description{
Calculate likelihood -- version 7
}
