% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_ij}
\alias{loglik_ij}
\title{Calculate likelihood of (beta_hat_1[j], beta_hat_2[j])) given
rho, b, sigma_1j, sigma_2j, s_1j, s_2j where
alpha_kj ~ N(0, sigma_kj)}
\usage{
loglik_ij(rho, g, gp, q, sigma1, sigma2, b1, b2, s1, s2)
}
\description{
Calculate likelihood of (beta_hat_1[j], beta_hat_2[j])) given
rho, b, sigma_1j, sigma_2j, s_1j, s_2j where
alpha_kj ~ N(0, sigma_kj)
}
