% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_mat}
\alias{loglik_mat}
\title{Calculate likelihood matrix -- version 7}
\usage{
loglik_mat(
  rho,
  g,
  gp,
  q,
  tsigma1,
  tsigma2,
  tbeta_hat_1,
  tbeta_hat_2,
  tseb1,
  tseb2
)
}
\value{
A matrix that is p by K where p is the number of SNPs
and K is the grid size
}
\description{
Calculate likelihood matrix -- version 7
}
