% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_samps_Z0}
\alias{loglik_samps_Z0}
\title{Calculate likelihood matrix for LOO (SNPs by posterior samples)}
\usage{
loglik_samps_Z0(tg, tgp, tq, rho, tsigma1, tsigma2, tpi, tbeta_hat_1,
  tbeta_hat_2, tseb1, tseb2)
}
\value{
A matrix that is p by L where p is the number of SNPs
and L is the number of posterior samples
}
\description{
Calculate likelihood matrix for LOO (SNPs by posterior samples)
}
