% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_pi_rho.R
\name{map_pi_rho}
\alias{map_pi_rho}
\title{Estimate pi and mixture proportions under null model}
\usage{
map_pi_rho(X, mix_grid, rho_start = 0, tol = 1e-07, n.iter = 20,
  null_wt = 10, z_prior_func = function(z) {     dnorm(z, 0, 0.5, log =
  TRUE) }, optmethod = c("mixSQP", "mixIP"), control = list(),
  warm = FALSE)
}
\arguments{
\item{X}{An object of class cause_data containing data for the two traits.}

\item{mix_grid}{An object of class cause_grid containing variance pair candidates}

\item{rho_start}{Starting value for rho}

\item{null_wt}{Specifies the prior weight on the first entry of grid}

\item{z_prior_func}{Prior function for z = arctanh(rho)}
}
\description{
This function is called by est_cause_params and generally does not need to be
directly accessed.
Estimate the MAP for rho and the mixing proportions using
coordinate descent. Causal effect (gamma) and the effect of U (eta) are fixed at zero.
}
