% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_format_cause.R
\name{new_cause_data}
\alias{new_cause_data}
\title{Create a new cause_data object}
\usage{
new_cause_data(x = data.frame())
}
\arguments{
\item{x}{a data.frame that includes columns snp, beta_hat_1, seb1, beta_hat_2, seb2 in any order.
x may also contain other columns}
}
\value{
and object of class cause_data and data.frame.
}
\description{
Create a new cause_data object
}
