% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R
\name{new_cause_data_fit}
\alias{new_cause_data_fit}
\title{Create a new cause_data_fit object}
\usage{
new_cause_data_fit(x = data.frame())
}
\arguments{
\item{x}{a data.frame that includes columns snp, beta_hat_1, seb1, beta_hat_2, seb2, delta_elpd,
prob_Z1_sharing, prob_Z1_causal in any order.
x may also contain other columns}
}
\value{
and object of class cause_data and data.frame.
}
\description{
Create a new cause_data_fit object
}
