% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot.cause}
\alias{plot.cause}
\title{Plot CAUSE}
\usage{
\method{plot}{cause}(res, intern = FALSE)
}
\arguments{
\item{res}{object of class cause}
}
\description{
Plot posteriors for confounding and causal models, display summary tables
}
