% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot.cause_post}
\alias{plot.cause_post}
\title{Plot posteriors for one CAUSE fit}
\usage{
plot.cause_post(fit)
}
\arguments{
\item{fit}{object of class cause_post `}
}
\description{
Plot posteriors for one CAUSE fit
}
