% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.cause_summary}
\alias{print.cause_summary}
\title{Print CAUSE Summary}
\usage{
\method{print}{cause_summary}(x, digits = 2)
}
\arguments{
\item{x}{object of class cause_summary}

\item{digits}{significant digits to report.}
}
\description{
Print a CAUSE summary
}
