% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary_cause_post}
\alias{print.summary_cause_post}
\title{Print CAUSE fit Summary}
\usage{
\method{print}{summary_cause_post}(x)
}
\arguments{
\item{x}{object of class cause_summary_post}
}
\description{
Print a CAUSE fit summary
}
