% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_sample_elpd_loo.R
\name{recompute_elpd_table}
\alias{recompute_elpd_table}
\title{Recompute elpd table for a CAUSE fit that is already computed}
\usage{
recompute_elpd_table(res)
}
\arguments{
\item{res}{A cause fit from running the cause function}
}
\value{
A model table equivalent to res$elpd
}
\description{
Recompute elpd table for a CAUSE fit that is already computed
}
