% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gwas_pairs.R
\name{setup_gwas_pairs}
\alias{setup_gwas_pairs}
\title{Set up analysis directory for pairs of GWAS summary statistics}
\usage{
setup_gwas_pairs()
}
\description{
A helper function that prepares a working directory for running an analysis with CAUSE.
}
\details{
This function prepares a working directory to replicate the results in Section 2.3 of Morrison et al 2019
(https://www.biorxiv.org/content/10.1101/682237v3). For more details see https://jean997.github.io/cause/gwas_pairs.html.
}
