% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_quantile.R
\name{step_quantile}
\alias{step_quantile}
\title{Quantile of a step pdf}
\usage{
step_quantile(x, start, stop, p)
}
\arguments{
\item{x}{quantile or vector of quantiles}

\item{start}{vector of step starts}

\item{stop}{vector of step stops}

\item{p}{vector of step probabilities}
}
\description{
Quantile of a step pdf
}
