% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{summary.cause}
\alias{summary.cause}
\title{Summarize CAUSE Results}
\usage{
\method{summary}{cause}(res, ci_size = 0.95, digits = 2)
}
\arguments{
\item{res}{object of class cause.}

\item{ci_size}{size of posterior credible intervals.}

\item{digits}{significant digits to report.}
}
\description{
Summarize p-value testing that the causal model fits the data better and the posterior
medians and credible intervals for parameters
}
