% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cause_post}
\alias{summary.cause_post}
\title{Summarize CAUSE Results for a single fit}
\usage{
\method{summary}{cause_post}(fit, ci_size = 0.95, digits = 2)
}
\arguments{
\item{fit}{object of class cause_post.}

\item{ci_size}{size of posterior credible intervals.}

\item{digits}{significant digits to report.}
}
\description{
Summarize posterior
medians and credible intervals for parameters
}
