% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R, R/package.R
\docType{package}
\name{cause}
\alias{cause}
\alias{cause-package}
\title{CAUSE}
\usage{
cause(X, variants, param_ests, sigma_g = 0.6, qalpha = 1, qbeta = 10,
  max_q = 1)
}
\arguments{
\item{X}{An object of class cause_data containing data for the two traits.}

\item{param_ests}{Object of class cause_params output by est_cause_params.
This contains estimates of the mixing proportions and an estimate of rho,
the correlation in test statistics that is due to overlapping samples or population structure.}

\item{sigma_g}{Parameter specifying the prior distribution of gamma and eta.
gamma ~ N(0, sigma_g), eta ~ N(0, sigma_g).}

\item{qalpha, qbeta}{Parameters defining the prior distribution of q.
q ~ Beta(qalpha, qbeta)}
}
\value{
A list with items conf, full, elpd, summary, and plot.
}
\description{
Fit CAUSE confounding only and full (causal) models,
calculate ELPD test statistic and estimate posteriors

Implementation of CAUSE described in Morrison et al. (2018)
}
\author{
Jean Morrison <jeanm@uchicago.edu>
}
