% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause.R, R/package.R
\docType{package}
\name{cause}
\alias{cause}
\title{CAUSE}
\usage{
cause(
  X,
  param_ests,
  variants = X$snp,
  sigma_g,
  qalpha = 1,
  qbeta = 10,
  max_q = 1,
  force = FALSE,
  n_start_gamma_eta = 21,
  n_start_q = 11
)
}
\arguments{
\item{X}{An object of class cause_data containing data for the two traits.}

\item{param_ests}{Object of class cause_params output by est_cause_params.
This contains estimates of the mixing proportions and an estimate of rho,
the correlation in test statistics that is due to overlapping samples or population structure.}

\item{variants}{A vector of variants to include in the analysis.}

\item{sigma_g}{Parameter specifying the prior distribution of gamma and eta.
gamma ~ N(0, sigma_g), eta ~ N(0, sigma_g).}

\item{qalpha, qbeta}{Parameters defining the prior distribution of q.
q ~ Beta(qalpha, qbeta)}

\item{max_q}{Largest value of q to be allowed. If max_q < 1 then the prior will be truncated.}

\item{force}{If true, do not give an error if parameter estimates did not converge.}

\item{n_start_gamma_eta, n_start_q}{Number of starting bins for grid approximation.
You shouldn't need to change these but if you are suspicious about your results, you might try increasing them.
It's best to use odd numbers.}
}
\value{
An object of class "cause" that contains posterior estimates for causal and sharing models as well as
results of the model comparison test. See Details.
}
\description{
Fit CAUSE sharing and causal models,
calculate ELPD test statistic and estimate posteriors

Implementation of CAUSE described in Morrison et al. (2020)
}
\details{
This function estimates posterior distributions for gamma, eta, and q under the sharing and causal models
and computes a test statistic comparing the two models. The returned object contains

sharing, causal: posterior estimates under model.

elpd: A data frame giving estimated difference in elpd between pairs of sharing, causal and null model with standard errors.
The null model is the model with gamma = eta = q = 0 and allows neither a causal effect nor correlated pleiotropy.

data: The data used to compute the object. This data frame also contains posterior estimates for each variant of acting
through U under both models.

sigma_g: Prior variance of gamma and eta. This value is chosen based on the data by default but can be supplied using the
sigma_g parameter in the function call.

qalpha, qbeta: The prior distribution of q is Beta(qalpha, qbeta)

Functions summary() and plot() can be used with cause objects. See ?summary.cause and ?plot.cause
}
\author{
Jean Morrison <jeanm@uchicago.edu>
}
