% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cause_post}
\alias{plot.cause_post}
\title{Plot posteriors for one CAUSE fit}
\usage{
\method{plot}{cause_post}(fit, type = c("posteriors", "data"), data = NULL, pval_thresh = 5e-08)
}
\arguments{
\item{fit}{object of class cause_post `}

\item{type}{Either "posteriors" or "data". See details.}

\item{data}{If type=="data" then an object of type cause_data_fit must be supplied}

\item{pval_thresh}{p-value threshold used when type=="data".}
}
\description{
Plot posteriors for one CAUSE fit
}
\details{
If type == "posteriors", the function plots the marginal posterior
distribution of each parameter. If type=="data", the function plots data colored
by the probability that Z = 1.
}
