% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_from_grid.R
\name{samp_from_grid}
\alias{samp_from_grid}
\title{Sample from grid approximation of posterior distribuitions}
\usage{
samp_from_grid(post, params, nsamps)
}
\arguments{
\item{params}{Vector of names of parameters to sample}

\item{joint_post}{Data frame giving the joint posterior}
}
\description{
Sample from the joint posterior of parameters
}
