% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_cause_pipeline.R
\name{setup_cause_pipeline}
\alias{setup_cause_pipeline}
\title{Set up analysis directory for pairs of GWAS summary statistics}
\usage{
setup_cause_pipeline(download_ld = FALSE, download_eur_ld_scores = FALSE)
}
\description{
A helper function that prepares a working directory for running an analysis with CAUSE.
}
\details{
This function downloads Snakemake pipeline and R code needed to run CAUSE on many pairs of traits.
}
